refactor(globals): split commands/globals.rs into module dir
617-line file split into mod/helpers/collectors/render/tests for
maintainability. Public API preserved (show, collect, GlobalCliEntry,
GlobalsReport unchanged).

Also pins sysinfo to 0.38 (0.39 requires rustc 1.95).
