+++ Cargo.toml
@@ -41,7 +41,7 @@ sha2 = "0.11"
 dirs = "6"
 
 # 系统信息（磁盘空间检查）
-sysinfo = "0.39"
+sysinfo = "0.38"
 
 # 错误处理
 thiserror = "2"
